MergeTree: Organize and Deprecated Collections
