+++ packages/dds/merge-tree/src/partialLengths.ts
@@ -4,8 +4,7 @@
  */
 
 import { assert } from "@fluidframework/common-utils";
-import { Property } from "./base";
-import { RedBlackTree } from "./collections";
+import { Property, RedBlackTree } from "./collections";
 import { UnassignedSequenceNumber } from "./constants";
 import {
     CollaborationWindow,
